feat: tab bar design complete
