+++ app/(tabs)/(my)/index.tsx
@@ -1,4 +1,4 @@
-import TabScreen from '@/components/CustomTab/TabScreen';
+import TabScreen from '@/components/molecules/CustomTab/TabScreen';
 import { ManageStudy, JoinedStudy, InterestStudy } from '@/components/organisms/MyPage/Main/StudyTabs';
 import FormalInfo from '@/components/organisms/MyPage/Main/FormalInfo';
